requirements.txt
pandas
babel
mysql-connector-python==8.0.33
openpyxl
sqlalchemy
psycopg2
requests
datetime
matplotlib
python-telegram-bot==20.0
nest_asyncio
asyncio
